Struct owned_alloc::UninitAlloc

pub struct UninitAlloc<T>where
    T: ?Sized,{ /* private fields */ }

Dynamic allocation of a T whose memory is considered uninitialized. The allocation is freed on drop. If the size of the allocation is zero, no allocation is performed and a dangling pointer is used (just like in std). For the drop checker, the type acts as if it contains a T due to usage of PhantomData<T>.

Implementations

impl<T> UninitAlloc<T>

pub fn new() -> Self

Creates room for a T. In case of allocation error, the handler registered via stdlib is called.

pub fn try_new() -> Result<Self, AllocErr>

Creates room for a T. In case of allocation error, Err is returned.

pub fn init(self, val: T) -> OwnedAlloc<T>

Initializes the memory and returns the allocation now considered initialized.

impl<T> UninitAlloc<T>where T: ?Sized,

pub unsafe fn init_in_place<F>(self, init: F) -> OwnedAlloc<T>where F: FnOnce(&mut T),

Calls a function with a mutable reference to uninitialized memory and returns the allocation now considered initialized. The passed function is expected to initialize the memory.

Safety

This function is unsafe because the passed function might not initialize the memory correctly.

pub unsafe fn from_raw(nnptr: NonNull<T>) -> Self

Recreate the UninitAlloc from a raw non-null pointer.

Safety

This functions is unsafe because passing the wrong pointer leads to undefined behaviour.

pub fn raw(&self) -> NonNull<T>

Returns the raw non-null pointer of the allocation.

pub fn into_raw(self) -> NonNull<T>

“Forgets” dropping the allocation and returns its raw non-null pointer.